==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Erica Sawyer"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity/")
```


# Setting the Environment
```{r set env}
# Set Seed
set.seed(238428)

```

# Load Libraries
```{r lib}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq, install=FALSE)

```
# Load in Data
```{r data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

#intuition check
min(sample_sums(raw_preprocessed_physeq))

# Make metadata dataframe
metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

#Setting colors for stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")


```


# Goals
1. Calculate the Hill Diversity of the samples. 
2. Evaluate the rarefaction curves. 
3. Evaluate the Diversity values. 
4. Makes notes of specific samples and their seq depth. 

# Diversity Calculations with iNEXT 

```{r calc-div}
# prepare input data 
iNEXT_input_df <- 
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()
# Quick check
dim(iNEXT_input_df)

# Run iNEXT: Calculate the Hill Numbers 
# Note that: Species in ROWS, Samples in COLUMNS 
# Remember to set the seed! 
#iNEXT_data <- iNEXT(iNEXT_input_df, 
#                    q = c(0,1,2), datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the Diversity! 
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
```

# Plot Diversity 
```{r plot-rarefaction}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)


# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
```

# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) + 
  theme(legend.position = "none")
```

# Manually Plot Diversity

## Rarefaction Curve
```{r iNEXT manual plot}
iNEXT_manual_df <-
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  #fix sample names
  mutate(names = gsub(names, pattern= "[.]", replace = "-"),
         names = gsub(names, pattern= "X", replace= "")) %>%
  # join with metadata
  left_join(., metadata_df, by = "names") %>%
  # add colors to df
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)),
            by = "station") #%>%
  #head()
  #View()

# Inspect
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

#Plot it
iNEXT_manual_df %>%
  # filter out rows that are calculated by rarefaction
  dplyr::filter(Method == "Rarefaction") %>%
  # make plot
  ggplot(aes(x=m, y=qD, color = station, group=names)) +
  # line
  geom_line() +
  #geom_point() +
  facet_grid(~Order.q~station, scales="fixed") +
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")
  

```


# Environmental Variable Check

```{r environmental-pairs, fig.width=8, fig.height=8}
head(metadata_df)
# Pull out environmental 
env_df <- 
  metadata_df %>%
  dplyr::select(names, water_tempC:DO_mgL)
# inspect
head(env_df)

# plot correlation
pairs(dplyr::select(env_df, -names), upper.panel = NULL)

```

# Diversity vs Salinity
```{r div-vs-salinity, fig.width=8, fig.height=8}
#PSU = practical salinity unit; 1 PSU= 1 g salt per 1000 g h2o (1 ppt)
#Freshwater = 0-0.5
#Brackish = 0.5-30
#Marine= >30

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x=salinity_psu, y=qD)) +
  facet_wrap(~Order.q, scales="free") +
  geom_point(aes(color = station)) +
  stat_smooth(method = "lm", formula = y ~poly(x, 2)) + #something is wrong here
  labs(x = "Salinity (PSU", y = "# of ASVs") +
  scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")


```

# Session Information
```{r session-info}
devtools::session_info()

```
